// packages/frontend/src/types/documentation.ts
export type UserRole =
  | "developer"
  | "salon_owner"
  | "salon_employee"
  | "salon_customer"
  | "system_admin"
  | "guest";